Upgrade Spring Boot 3.1.0

+++ src/main/java/com/jvbarbosa/dscommerce/services/ProductService.java
@@ -8,16 +8,14 @@ import com.jvbarbosa.dscommerce.entities.Product;
 import com.jvbarbosa.dscommerce.repositories.ProductRepository;
 import com.jvbarbosa.dscommerce.services.exceptions.DatabaseException;
 import com.jvbarbosa.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import javax.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -27,9 +25,9 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product result = repository.findById(id).orElseThrow(
+        Product product = repository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("Recurso não encontrado"));
-        return new ProductDTO(result);
+        return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
@@ -43,7 +41,6 @@ public class ProductService {
         Product entity = new Product();
         copyDtoToEntity(dto, entity);
         entity = repository.save(entity);
-
         return new ProductDTO(entity);
     }
 
@@ -53,23 +50,21 @@ public class ProductService {
             Product entity = repository.getReferenceById(id);
             copyDtoToEntity(dto, entity);
             entity = repository.save(entity);
-
             return new ProductDTO(entity);
         }
         catch (EntityNotFoundException e) {
             throw new ResourceNotFoundException("Recurso não encontrado");
         }
-
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        try {
-            repository.deleteById(id);
-        }
-        catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Recurso não encontrado");
-        }
+    	if (!repository.existsById(id)) {
+    		throw new ResourceNotFoundException("Recurso não encontrado");
+    	}
+    	try {
+            repository.deleteById(id);    		
+    	}
         catch (DataIntegrityViolationException e) {
             throw new DatabaseException("Falha de integridade referencial");
         }
@@ -80,12 +75,12 @@ public class ProductService {
         entity.setDescription(dto.getDescription());
         entity.setPrice(dto.getPrice());
         entity.setImgUrl(dto.getImgUrl());
-
+        
         entity.getCategories().clear();
         for (CategoryDTO catDto : dto.getCategories()) {
-            Category cat = new Category();
-            cat.setId(catDto.getId());
-            entity.getCategories().add(cat);
+        	Category cat = new Category();
+        	cat.setId(catDto.getId());
+        	entity.getCategories().add(cat);
         }
     }
 }
